backend/requirements.txt
alembic==1.13.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
email-validator==2.3.0
fastapi==0.121.0
h11==0.16.0
httpx==0.26.0
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pytest==8.0.0
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
svix==1.16.0
